# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PhotoBomb API
  version: 1.0.0
  description: |
    Production-ready photo service API supporting uploads, metadata management, 
    search, albums, and sharing with presigned URLs and async processing.
  contact:
    name: PhotoBomb Team
    email: api@photobomb.example.com
servers:
  - url: https://api.photobomb.example.com/api/v1
    description: Production
  - url: https://api-staging.photobomb.example.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /albums:
    get:
      tags: [albums]
      summary: List user's albums
      responses:
        '200':
          description: Album list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'

    post:
      tags: [albums]
      summary: Create album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  example: "Summer Vacation 2024"
                description:
                  type: string
                photo_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
  /albums/{album_id}:
    get:
      tags: [albums]
      summary: Get album details
      parameters:
        - name: album_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Album with photos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumDetailed'

    patch:
      tags: [albums]
      summary: Update album
      parameters:
        - name: album_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                add_photo_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                remove_photo_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Album updated

    delete:
      tags: [albums]
      summary: Delete album
      parameters:
        - name: album_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Album deleted (photos preserved)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
  schemas:
...
    Photo:
      type: object
      properties:
        photo_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        filename:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        sha256:
          type: string
        taken_at:
          type: string
          format: date-time
          nullable: true
          description: From EXIF DateTimeOriginal
        uploaded_at:
          type: string
          format: date-time
        thumb_urls:
          type: object
          properties:
            thumb_256:
              type: string
              format: uri
            thumb_512:
              type: string
              format: uri
            thumb_1024:
              type: string
              format: uri
        caption:
          type: string
          nullable: true
        favorite:
          type: boolean
        archived:
          type: boolean
...
    Album:
      type: object
      properties:
        album_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        cover_photo_id:
          type: string
          format: uuid
          nullable: true
        photo_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AlbumDetailed:
      allOf:
        - $ref: '#/components/schemas/Album'
        - type: object
          properties:
            photos:
              type: array
              items:
                $ref: '#/components/schemas/Photo'
...